pyproject.toml
[project]
name = "faststring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-benchmark>=5.1.0",
]
